# backend/src/wishes/dto/update-wish.dto.ts
import { IsNumber, IsOptional, IsString, IsUrl } from "class-validator";

export class UpdateWishDto {
  @IsString()
  @IsOptional()
  name?: string;
  @IsUrl()
  @IsOptional()
  link?: string;
  @IsUrl()
  @IsOptional()
  image?: string;
  @IsNumber()
  @IsOptional()
  price?: number;
  @IsString()
  @IsOptional()
  description?: string;
}